requirements.txt
Django==4.2.11
python-dotenv==1.0.1
djangorestframework==3.15.1
mysqlclient==2.1.1
redis==5.0.4
djangorestframework-simplejwt==5.3.1